+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 --------------------
 * Updated `sw_f107` to allow reading old and new `historic` data file format
 * Added exampled to the documentation for F10.7 and Kp methods
+* Removed warning for not using the deprecated `freq` kwarg
+* Removed deprecated function `load_csv_data`
 
 [0.0.6] - 2022-07-08
 --------------------